Add files via upload
📌 Project Overview
This project is a structured data analysis of a music store using PostgreSQL. It is inspired by a guided tutorial on YouTube, with additional improvements and insights. The project explores customer purchases, sales trends, employee performance, and popular albums/tracks through SQL queries.

🎓 Credits & Acknowledgments
This project is based on a YouTube tutorial by Rishabh Mishra. I have optimized few queries of the analysis to enhance my SQL and data analysis skills.